services/rag-service/pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
description = "RAG Microservice for document ingestion and context retrieval"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docling>=2.55.0",
    "docling-core>=2.11.0",
    "ollama>=0.6.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "minio>=7.2.0",
    # OCR and document parsing dependencies
    "rapidocr>=1.4.0",
    "onnxruntime>=1.17.0",
]

[project.optional-dependencies]
# CPU-only installation (for Docker and servers without GPU)
cpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

# GPU installation (for systems with CUDA)
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
